edit the setting of nginx
make sure all the request that get from the port 80 that received by the nginx is sent to the port of the django server.
The upstream app is django which run on the local side of the backend.
Nginx listens to the port on 80 and send all the request from / to the local server